# PERF: Fix regression in datetime ops