@TomAugspurger 2017-10-25
This also seems to help with the `timeseries.SeriesArithmetic.time_add_offset` benchmarks:

head:

```
[ 33.33%] ··· Running timeseries.SeriesArithmetic.time_add_offset_delta                                                                               4.30ms
[ 66.67%] ··· Running timeseries.SeriesArithmetic.time_add_offset_fast                                                                                11.0ms
[100.00%] ··· Running timeseries.SeriesArithmetic.time_add_offset_slow                                                                                 703ms
```

master:

```
[ 33.33%] ··· Running timeseries.SeriesArithmetic.time_add_offset_delta                                                                                499ms
[ 66.67%] ··· Running timeseries.SeriesArithmetic.time_add_offset_fast                                                                                 489ms
[100.00%] ··· Running timeseries.SeriesArithmetic.time_add_offset_slow                                                                                 1.05s
```