@TomAugspurger 2017-10-25
Cool, this also fixes the slowdown in the stata benchmarks:

head:

```
[  0.00%] ·· Building for existing-py_Users_taugspurger_Envs_pandas-dev_bin_python3.6
[  0.00%] ·· Benchmarking existing-py_Users_taugspurger_Envs_pandas-dev_bin_python3.6
[ 25.00%] ··· Running packers.STATA.time_write_stata                                                                                                  33.2ms
[ 50.00%] ··· Running packers.STATA.time_write_stata_with_validation                                                                                  49.5ms
[ 75.00%] ··· Running packers.packers_read_stata.time_packers_read_stata                                                                              39.6ms
[100.00%] ··· Running packers.packers_read_stata_with_validation.time_packers_read_stata_with_validation                                              48.0ms
```

master:

```
[  0.00%] ·· Building for existing-py_Users_taugspurger_Envs_pandas-dev_bin_python3.6
[  0.00%] ·· Benchmarking existing-py_Users_taugspurger_Envs_pandas-dev_bin_python3.6
[ 25.00%] ··· Running packers.STATA.time_write_stata                                                                                                   502ms
[ 50.00%] ··· Running packers.STATA.time_write_stata_with_validation                                                                                   529ms
[ 75.00%] ··· Running packers.packers_read_stata.time_packers_read_stata                                                                               631ms
[100.00%] ··· Running packers.packers_read_stata_with_validation.time_packers_read_stata_with_validation                                               650ms
```