@jbrockmendel 2017-10-25
Are the `isinstance` checks being called zillions of times?  `ABCDateOffset` was introduced recently, and `isinstance(x, ABCFoo)` is about 2x slower than `isinstance(x, Foo)`