@TomAugspurger 2017-10-25
> Are the isinstance checks being called zillions of times?

On master they were being called once per row. The change here skips that, since if we have a `datetime64` or `timedelta64` array, then we know none of them are offsets.